Refactor the utils elements
