+++ kryptone/utils/urls.py
@@ -1,15 +1,9 @@
-from string import Template
 from collections import defaultdict
-import bisect
-import datetime
 import pathlib
-import random
 import re
 from functools import lru_cache
-from collections import OrderedDict
 from urllib.parse import urljoin, urlparse
 
-import pytz
 import requests
 
 from kryptone import logger
@@ -39,7 +33,7 @@ class URL:
         return self.raw_url == obj
 
     def __add__(self, obj):
-        return urljoin(self.raw_url, obj)
+        return URL(urljoin(self.raw_url, obj))
 
     def __contains__(self, obj):
         return obj in self.raw_url
@@ -68,10 +62,6 @@ class URL:
             self.raw_url.endswith('#')
         ])
 
-    @classmethod
-    def create(cls, url):
-        return cls(url)
-
     @property
     def is_file(self):
         path = settings.GLOBAL_KRYPTONE_PATH / 'data/file_extensions.txt'
@@ -99,6 +89,10 @@ class URL:
     def url_stem(self):
         return self.as_path.stem
 
+    @classmethod
+    def create(cls, url):
+        return cls(url)
+    
     def is_same_domain(self, url):
         incoming_url_object = urlparse(url)
         return incoming_url_object.netloc == self.url_object.netloc
@@ -181,43 +175,20 @@ class URL:
             return False
         return list(drop_while(clean_values, result))
 
-    def paginate(self, k=10, regex_path=None, param=None):
-        """Increase the pagination number provided
-        on a given url"""
-        if regex_path is None and param is None:
-            # If we have nothing, classically, page
-            # is used to paginate in urls
-            param = 'page'
-        return 1
-
-
-class CompareUrls:
-    """Check the similarity between two different urls
-
-    >>> TestUrl('https://example.com', 'http://example.com/')
-    ... True
-    """
-
-    def __init__(self, current_url, url_to_test):
-        if isinstance(current_url, str):
-            current_url = URL(current_url)
-
-        if isinstance(url_to_test, str):
-            url_to_test = URL(url_to_test)
-
-        self.current_url = current_url
-        self.url_to_test = url_to_test
-        self.test_result = self.current_url.compare(url_to_test)
-
-    def __repr__(self):
-        return f'<TestUrl: result={self.test_result}>'
-
-    def __bool__(self):
-        return self.test_result
+    # def paginate(self, k=10, regex_path=None, param=None):
+    #     """Increase the pagination number provided
+    #     on a given url"""
+    #     if regex_path is None and param is None:
+    #         # If we have nothing, classically, page
+    #         # is used to paginate in urls
+    #         param = 'page'
+    #     return 1
 
 
 class BaseURLTestsMixin:
-    blacklist = []
+    blacklist = set()
+    blacklist_distribution = defaultdict(list)
+    error_message = "{url} was blacklisted by filter '{filter_name}'"
 
     def __call__(self, url):
         pass
@@ -229,25 +200,21 @@ class BaseURLTestsMixin:
 
 
 class URLIgnoreTest(BaseURLTestsMixin):
-    """Excludes any url that match the provided paths
-    in the `paths` parameter
+    """Ignore every url in which the provided
+    paths match one or many sections of the url
 
-    >>> class Spider(BaseCrawler):
-            url_passes_tests = [
-                URLPassesTest(
-                    'simple_test',
-                    paths=[
-                        '/example'
-                    ]
-                )
-            ]
+    For example, `example.com/1` will be 
+    ignored with `/1`
     """
 
-    def __init__(self, name, *, paths=[], ignore_files=[]):
+    def __init__(self, name, *, paths=[]):
         self.name = name
+        if not isinstance(paths, (list, tuple)):
+            raise ValueError("'paths' should be a list or a tuple")
         self.paths = set(paths)
-        self.failed_paths = []
-        self.ignore_files = ignore_files
+
+    def __repr__(self):
+        return f'<{self.__class__.__name__} {self.paths}>'
 
     def __call__(self, url):
         exclusion_truth_array = []
@@ -259,279 +226,72 @@ class URLIgnoreTest(BaseURLTestsMixin):
         # others as False
         for path in self.paths:
             if path in url.url_object.path:
-                self.failed_paths.append(path)
+                self.blacklist.add(path)
                 exclusion_truth_array.append(True)
             else:
                 exclusion_truth_array.append(False)
 
-            # if url.is_file and self.ignore_files:
-            #     if url.get_extension in self.ignore_files:
-            #         result.append(False)
-
         if any(exclusion_truth_array):
-            logger.warning(f"{url} failed test '{self.name}'")
-            return False
-        return True
-
-    @lru_cache(maxsize=10)
-    def default_ignored_files(self):
-        path = settings.GLOBAL_KRYPTONE_PATH / 'data/file_extensions.txt'
-        sorted_values = sorted(read_document(path, as_list=True))
-        return list(drop_while(lambda x: x == '', sorted_values))
+            logger.warning(
+                self.error_message.format(
+                    url=url, 
+                    filter_name=self.name
+                )
+            )
+            return True
+        return False
 
 
 class URLIgnoreRegexTest(BaseURLTestsMixin):
-    """Checks if an url is able to pass a
-    a given test
+    """Ignore every url in which the provided
+    regex path match a specific section of the url
 
-    >>> class Spider(BaseCrawler):
-            url_passes_tests = [
-                URLPassesRegexTest(
-                    'simple_test',
-                    regex=r'\/a$'
-                )
-            ]
+    For example, `example.com/1` will be 
+    ignored with `\/\d+`
     """
 
-    def __init__(self, name, *, regex=None):
+    def __init__(self, name, *, regex):
         self.name = name
         self.regex = re.compile(regex)
-
-    def __call__(self, url):
-        result = self.regex.search(url)
-        if result:
-            return False
-        logger.warning(f"{url} failed test: '{self.name}'")
-        return True
-
-
-class URLPassesRegexTest(BaseURLTestsMixin):
-    """Only include and keep urls that successfully pass
-    the provided regex test
-    """
-
-    def __init__(self, name, *, regex=None):
-        self.name = name
-        self.regex = re.compile(regex)
-
-    def __call__(self, url):
-        result = self.regex.search(url)
-        if result:
-            logger.warning(f"{url} succeeded test: '{self.name}'")
-            return True
-        return False
-
-
-class Some:
-    """Receive the results of all tests and
-    if any one of them returns `True`, consider
-    the url to be valid for visit"""
-    
-    def __init__(self, *tests):
-        self.tests = tests
-
-    def __call__(self, url):
-        test_result = []
-        for test in self.tests:
-            test_result.append(test)
-
-        if any(test_result):
-            return True
-        return False
-
-
-class URLIterator:
-    _urls_to_visit = set()
-    _visited_urls = set()
-    _seen_urls = set()
-    _grouped_by_page = defaultdict(set)
-    _current_url = None
-
-    def __init__(self, start_urls=[], sort_urls=False):
-        self.sort_urls = sort_urls
-        result = self.pre_save(start_urls)
-        self._urls_to_visit.update(result)
 
     def __repr__(self):
-        name = self.__class__.__name__
-        return f'<{name} urls_to_visit={self.urls_to_visit_count} visited_urls={self.visited_urls_count}>'
+        return f'<{self.__class__.__name__} [{self.regex}]>'
 
-    def __iter__(self):
-        for url in self._urls_to_visit:
-            yield url
-
-    def __contains__(self, url):
-        return any([
-            str(url) in self._urls_to_visit,
-            str(url) in self._visited_urls
-        ])
-
-    def __len__(self):
-        """Returns the amount of urls
-        left to visit"""
-        return len(self._urls_to_visit)
-
-    def __getitem__(self, index):
-        url = list(self._urls_to_visit)[index]
-        return URL(url)
-
-    @property
-    def empty(self):
-        return len(self._urls_to_visit) == 0
-
-    @property
-    def urls_to_visit(self):
-        for url in self._urls_to_visit:
-            yield URL(url)
-
-    @property
-    def visited_urls(self):
-        for url in self._visited_urls:
-            yield URL(url)
-
-    @property
-    def urls_to_visit_count(self):
-        return len(self._urls_to_visit)
-
-    @property
-    def visited_urls_count(self):
-        return len(self._visited_urls)
-
-    @property
-    def total_urls_count(self):
-        return sum([self.urls_to_visit_count, self.visited_urls_count])
-
-    @property
-    def completion_rate(self):
-        try:
-            result = self.urls_to_visit_count / self.visited_urls_count
-            return round(result, 2)
-        except ZeroDivisionError:
-            return float(0)
-
-    @property
-    def next_url(self):
-        try:
-            return list(self.urls_to_visit)[0]
-        except IndexError:
-            return None
-
-    @property
-    def grouped_by_page(self):
-        container = OrderedDict()
-        for key, values in self._grouped_by_page.items():
-            container[key] = list(values)
-        return container
-
-    def pre_save(self, urls):
-        # final_urls = set()
-        # urls = map(lambda x: URL(x), urls)
-        # for url in urls:
-        #     if url.is_file:
-        #         continue
-        #     final_urls.add(str(url))
-        # return list(final_urls)
-        return urls
-
-    def backup(self):
-        return {
-            'date': str(datetime.datetime.now(tz=pytz.UTC)),
-            'urls_to_visit': list(self._urls_to_visit),
-            'visited_urls': list(self._visited_urls),
-            'statistics': {
-                'last_visited_url': str(self._current_url) if self._current_url is not None else None,
-                'urls_to_visit_count': self.urls_to_visit_count,
-                'visited_urls_count': self.visited_urls_count,
-                'total_urls': sum([self.urls_to_visit_count, self.visited_urls_count]),
-                'completion_rate': self.completion_rate
-            }
-        }
-
-    def append(self, url):
-        self._seen_urls.add(url)
-
-        if url in self._urls_to_visit:
-            return False
-
-        if url in self._visited_urls:
-            return False
-
-        self._urls_to_visit.add(url)
-        if self.sort_urls:
-            self._urls_to_visit = set(sorted(self._urls_to_visit))
-            self._visited_urls = set(sorted(self._visited_urls))
-
-    def appendleft(self, url):
-        urls_to_visit = list(self._urls_to_visit)
-        urls_to_visit.insert(0, url)
-        self._urls_to_visit = set(urls_to_visit)
-
-    def clear(self):
-        self._urls_to_visit.clear()
-        self._visited_urls.clear()
-
-    def reverse(self):
-        container = []
-        for i in range(self.urls_to_visit_count, 0, -1):
-            try:
-                container.append(list(self._urls_to_visit)[i - 1])
-            except IndexError:
-                continue
-        self._urls_to_visit = set(container)
-
-    def update(self, urls, current_url=None):
-        keys = self._grouped_by_page.keys()
-        if keys:
-            key = current_url or list(keys)[-1] + 1
-        else:
-            key = current_url or 1
-
-        for url in urls:
-            self._grouped_by_page[key].add(url)
-            self.append(url)
-
-    def get(self):
-        url = self._urls_to_visit.pop()
-        self._current_url = URL(url)
-        self._visited_urls.add(url)
-        return self._current_url
-
-
-class URLGenerator:
-    """Generates a set of urls using a template
-
-    >>> generator = URLGenerator('http://example.com/$id', params={'id': 'number'}, k=2)
-    ... ['http://example.com/1', 'http://example.com/2']
-    """
-
-    def __init__(self, template, params={}, k=10, start=0):
-        self.base_template_url = Template(template)
-
-        new_params = []
-        base_params = [params for _ in range(k)]
-        for i, param in enumerate(base_params, start=start):
-            new_param = {}
-            for key, value in param.items():
-                if value == 'number':
-                    new_param[key] = i
-            new_params.append(new_param)
-
-        self.urls = []
-        for i in range(k):
-            try:
-                self.urls.append(
-                    self.base_template_url.substitute(new_params[i])
+    def __call__(self, url):
+        result = self.regex.search(url)
+        if result:
+            logger.warning(
+                self.error_message.format(
+                    url=url, 
+                    filter_name=self.name
                 )
-            except KeyError:
-                self.urls.append(template)
+            )
+            return True
+        return False
 
-    def __iter__(self):
-        for url in self.urls:
-            yield url
 
-    def __aiter__(self):
-        for url in self.urls:
-            yield url
+# class URLPassesRegexTest(BaseURLTestsMixin):
+#     """Only include and keep urls that successfully pass
+#     the provided regex test
+#     """
 
-    def __len__(self):
-        return len(self.urls)
+#     def __init__(self, name, regex):
+#         self.name = name
+#         self.regex = re.compile(regex)
+
+#     def __repr__(self):
+#         return f'<{self.__class__.__name__} [{self.regex}]>'
+
+#     def __call__(self, url):
+#         result = self.regex.search(url)
+#         if result:
+#             # Indicate to not ignore
+#             # the url
+#             return False
+#         logger.warning(
+#             self.error_message.format(
+#                 url=url, 
+#                 filter_name=self.name
+#             )
+#         )
+#         return True
